clean: ♻ リファクタリング #12
