+++ src/component/Home/context/StartYearContext.tsx
@@ -9,6 +9,7 @@ export const StartYearContext = createContext<ContextType>({
   startYear: 1960,
   setStartYear: () => {},
 })
+
 export const StartYearProvider = ({children}: {children: React.ReactNode}) => {
   const [startYear, setStartYear] = useState<number>(1960)
